[feat] KeyboardAwareContainer, YellowWideButton을 통한 로직 통일 & [bug] Android KeyboardAvoidingView 에러 해결 (#98)
