+++ fourtoon-cookie/src/pages/DiaryWritePage/WriteDoneButtonLayout.tsx
@@ -1,12 +1,9 @@
-import { Alert, KeyboardAvoidingView, Platform, StyleSheet } from "react-native"
-import { LocalDate } from "@js-joda/core";
+import { Alert } from "react-native"
 import { useState } from "react";
 import { NavigationProp, useNavigation } from "@react-navigation/native";
-import { OS } from "../../types/os"
 import { Diary, DiaryStatus } from "../../types/diary";
 import { useSelectedCharacterStore } from "../../hooks/store/selectedCharacter";
 import { RootStackParamList } from "../../types/routing";
-import Button from "../../components/common/Button";
 
 import { AccountStatus } from "../../types/account";
 import { useCreateDiary, useUpdateDiary } from "../../hooks/server/diary";
@@ -17,6 +14,8 @@ import { useTranslationWithParentName } from "../../hooks/locale";
 import { SelectedCharacterNotExistError } from "../../types/error/character/SelectedCharacterNotExistError";
 import { showSuccessToast } from "../../system/toast";
 import { useDiaryWritePageContext } from "./DiaryWritePageProvider";
+import KeyboardAwareContainer from "../../components/common/KeyboardAwareContainer";
+import YellowWideButton from "../../components/common/YellowWideButton";
 
 
 const WriteDoneButtonLayout = () => {
@@ -96,43 +95,13 @@ const WriteDoneButtonLayout = () => {
     });
 
     return (
-        <KeyboardAvoidingView 
-            style={styles.bottomContainer} 
-            enabled={true}
-            keyboardVerticalOffset={80}
-            behavior={(Platform.OS == OS.IOS) ? 'padding' : 'height'}
-        >
-            <Button
-                title={commonT('done')}
-                onPress={handleWriteDoneButtonPress}
-                style={{
-                    ...styles.nextButton, 
-                    backgroundColor: isNextButtonEnabled ? '#FFC426' : '#DDDDDD'
-                }}
-                textStyle={styles.nextButtonText}
+        <KeyboardAwareContainer>
+            <YellowWideButton
+                isNextAvailabe={isNextButtonEnabled}
+                onNextButtonClick={handleWriteDoneButtonPress}
             />
-        </KeyboardAvoidingView>
+        </KeyboardAwareContainer>
     )
 }
 
 export default WriteDoneButtonLayout;
-
-const styles = StyleSheet.create({
-    bottomContainer: {
-        position: 'absolute',
-        bottom: 20,
-        left: 20,
-        right: 20,
-      },
-    nextButton: {
-        width: '100%',
-        height: 60,
-        borderRadius: 16,
-        justifyContent: 'center',
-        alignItems: 'center',
-      },
-      nextButtonText: {
-        fontSize: 17,
-        fontWeight: '600'
-      }
-});
